refactored form fields from sign up and login screens into global component, added server error handling and better styling to login screen
